Refactoring file structure + code format
