+++ AvaloniaUIClient/Converters/HalfSizeConverter.cs
@@ -14,5 +14,4 @@ internal sealed class HalfSizeConverter : IValueConverter
     
     public object ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture) =>
         throw new NotSupportedException();
-  
 }
