+++ AvaloniaUIClient/Infrastructure/DataTemplates/LogsDataTemplates.axaml
@@ -6,22 +6,38 @@
     <!--  Add Resources Here  -->
 
     <!--  Default log  -->
-    <DataTemplate x:Key="DevLogsDataTemplate" DataType="events:LogEvent">
-        <StackPanel VerticalAlignment="Center" Orientation="Vertical">
+    <DataTemplate
+        x:Key="DevLogsDataTemplate"
+        DataType="events:LogEvent">
+        <StackPanel
+            VerticalAlignment="Center"
+            Orientation="Vertical">
 
-            <StackPanel ToolTip.Tip="Log level" Orientation="Horizontal">
+            <StackPanel
+                ToolTip.Tip="Log level"
+                Orientation="Horizontal">
                 <avalonia:MaterialIcon Kind="Information" />
-                <TextBlock Margin="5,0,0,0" Text="{Binding Level}" />
+                <TextBlock
+                    Margin="5,0,0,0"
+                    Text="{Binding Level}" />
             </StackPanel>
 
-            <StackPanel ToolTip.Tip="TimeStamp" Orientation="Horizontal">
+            <StackPanel
+                ToolTip.Tip="TimeStamp"
+                Orientation="Horizontal">
                 <avalonia:MaterialIcon Kind="Clock" />
-                <TextBlock Margin="5,0,0,0" Text="{Binding Timestamp.Now}" />
+                <TextBlock
+                    Margin="5,0,0,0"
+                    Text="{Binding Timestamp.Now}" />
             </StackPanel>
 
-            <StackPanel ToolTip.Tip="Text" Orientation="Horizontal">
+            <StackPanel
+                ToolTip.Tip="Text"
+                Orientation="Horizontal">
                 <avalonia:MaterialIcon Kind="CommentText" />
-                <TextBlock Margin="5,0,0,0" Text="{Binding MessageTemplate.Text}" />
+                <TextBlock
+                    Margin="5,0,0,0"
+                    Text="{Binding MessageTemplate.Text}" />
             </StackPanel>
 
         </StackPanel>
