+++ AvaloniaUIClient/Infrastructure/IOC/IocWorker.cs
@@ -7,5 +7,4 @@ internal static class IocWorker
     public static IServiceCollection RegistredServies() =>
         new ServiceCollection()
             .WindowRegistrator();
-
 }
