+++ AvaloniaUIClient/Infrastructure/Views/DevPanelViews/LogsPanel/LogsDevView.axaml
@@ -72,7 +72,9 @@
                         <!--  Main|Logs  -->
                         <styles:ColorZone Grid.Row="1">
 
-                            <Grid RowDefinitions="25,*" ColumnDefinitions="*,*">
+                            <Grid
+                                RowDefinitions="25,*"
+                                ColumnDefinitions="*,*">
 
                                 <TextBlock
                                     Grid.Row="0"
@@ -128,7 +130,9 @@
 
                     <!--  Options|Logs categories  -->
                     <styles:ColorZone>
-                        <Grid RowDefinitions="25,*" ColumnDefinitions="*,*">
+                        <Grid
+                            RowDefinitions="25,*"
+                            ColumnDefinitions="*,*">
                             <Button
                                 Grid.Row="0"
                                 Grid.Column="1"
@@ -193,7 +197,9 @@
                             Classes="Body1 Head"
                             Text="Logger test" />
 
-                        <ListBox Grid.Row="1" Items="{Binding LoggerTests}" />
+                        <ListBox
+                            Grid.Row="1"
+                            Items="{Binding LoggerTests}" />
                     </Grid>
                 </styles:ColorZone>
             </TabItem>
