+++ AvaloniaUIClient/Views/TitleViews/HomeView.axaml
@@ -1,26 +1,28 @@
 <UserControl
-    d:DesignHeight="450"
-    d:DesignWidth="800"
-    mc:Ignorable="d"
     x:Class="AvaloniaUIClient.Views.TitleViews.HomeView"
     xmlns="https://github.com/avaloniaui"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:avalonia="clr-namespace:Material.Icons.Avalonia;assembly=Material.Icons.Avalonia"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+    d:DesignHeight="450"
+    d:DesignWidth="800"
+    mc:Ignorable="d">
 
-    <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center">
+    <StackPanel
+        HorizontalAlignment="Center"
+        VerticalAlignment="Center">
         <avalonia:MaterialIcon
+            Width="150"
             Height="150"
             HorizontalAlignment="Center"
-            Kind="Home"
             VerticalAlignment="Center"
-            Width="150" />
+            Kind="Home" />
         <TextBlock
-            Classes="Headline6"
             HorizontalAlignment="Center"
-            Text="Why am i here?"
-            VerticalAlignment="Center" />
+            VerticalAlignment="Center"
+            Classes="Headline6"
+            Text="Why am i here?" />
     </StackPanel>
 
 </UserControl>
