+++ Core/Infrastructure/Logging/InfoToLogSink.cs
@@ -7,12 +7,10 @@ namespace Core.Infrastructure.Logging;
 
 internal sealed class InfoToLogSink : ILogEventSink
 {
-    
     private readonly ICollectionRepository<ObservableCollection<LogEvent>,LogEvent>  _logStore;
 
     public InfoToLogSink(ICollectionRepository<ObservableCollection<LogEvent>, LogEvent> logStore) =>
         _logStore = logStore;
    
     public void Emit(LogEvent logEvent) => _logStore.AddIntoEnumerable(logEvent);
-    
 }
