+++ Core/Infrastructure/Services/NavigationService/BaseTypeNavigationService.cs
@@ -1,6 +1,6 @@
 using AppInfrastructure.Services.StoreServices.Parametrize;
 using AppInfrastructure.Stores.DefaultStore;
-using Core.Infrastructure.VMD;
+using Core.Infrastructure.VMD.Interfaces;
 
 namespace Core.Infrastructure.Services.NavigationService;
