+++ Core/Infrastructure/Services/NavigationService/BaseVmdNavigationService.cs
@@ -1,6 +1,6 @@
 using AppInfrastructure.Stores.DefaultStore;
 using Core.Infrastructure.Hosting;
-using Core.Infrastructure.VMD;
+using Core.Infrastructure.VMD.Interfaces;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Core.Infrastructure.Services.NavigationService;
@@ -9,8 +9,5 @@ public abstract class BaseVmdNavigationService<TBaseVMd> : BaseTypeNavigationSer
 {
     public BaseVmdNavigationService(IStore<TBaseVMd> store) 
         : base(store, 
-            type =>(TBaseVMd)HostWorker.Services.GetRequiredService(type))
-    {
-    }
-    
+            type =>(TBaseVMd)HostWorker.Services.GetRequiredService(type)) { }
 }
