+++ Core/Infrastructure/Services/ParseService/IParseService.cs
@@ -5,5 +5,4 @@ public interface IParseService
     public string Serialize(object nonSerialized);
 
     public T? DeSerialize<T>(string serialized);
-
 }
