+++ Core/Infrastructure/Stores/BaseVmdStore.cs
@@ -1,9 +1,6 @@
 using AppInfrastructure.Stores.DefaultStore;
-using Core.Infrastructure.VMD;
+using Core.Infrastructure.VMD.Interfaces;
 
 namespace Core.Infrastructure.Stores;
 
-public abstract class BaseVmdStore<TBaseVmd> : BaseLazyStore<TBaseVmd> where TBaseVmd : IBaseVmd
-{
-    
-}
+public abstract class BaseVmdStore<TBaseVmd> : BaseLazyStore<TBaseVmd> where TBaseVmd : IBaseVmd { }
