+++ Core/Infrastructure/Stores/Interfaces/TitleVmdStore.cs
@@ -1,0 +1,8 @@
+using Core.Infrastructure.VMD.Interfaces;
+
+namespace Core.Infrastructure.Stores.Interfaces;
+
+public sealed class TitleVmdStore : BaseVmdStore<ITitleVmd>
+{
+    
+}
