--- Core/Infrastructure/Stores/TitleVmdStore.cs
@@ -1,8 +1,0 @@
-using Core.Infrastructure.VMD;
-
-namespace Core.Infrastructure.Stores;
-
-public sealed class TitleVmdStore : BaseVmdStore<ITitleVmd>
-{
-    
-}
