+++ Core/Infrastructure/VMD/BaseMenuVmd.cs
@@ -1,11 +1,12 @@
+using Core.Infrastructure.VMD.Interfaces;
 using ReactiveUI;
 
 namespace Core.Infrastructure.VMD;
 
 public abstract class BaseMenuVmd<T> : BaseVmd,IMenuVmd<T>
 {
-    public IEnumerable<T> MenuItems { get; init; }
+    public IEnumerable<T>? MenuItems { get; init; }
 
-    protected IReactiveCommand? _navigationCommand;
+    protected IReactiveCommand? NavigationCommand;
     
 }
