+++ Core/Infrastructure/VMD/BaseVmd.cs
@@ -1,8 +1,6 @@
+using Core.Infrastructure.VMD.Interfaces;
 using ReactiveUI;
 
 namespace Core.Infrastructure.VMD;
 
-public class BaseVmd : ReactiveObject, IBaseVmd
-{
-    
-}
+public class BaseVmd : ReactiveObject, IBaseVmd { }
