--- Core/Infrastructure/VMD/IBaseVmd.cs
@@ -1,8 +1,0 @@
-using ReactiveUI;
-
-namespace Core.Infrastructure.VMD;
-
-public interface IBaseVmd: IReactiveObject
-{
-    
-}
