--- Core/Infrastructure/VMD/IMenuVmd.cs
@@ -1,7 +1,0 @@
-namespace Core.Infrastructure.VMD;
-
-public interface IMenuVmd<T> : IBaseVmd
-{
-    public IEnumerable<T> MenuItems { get; init; }
-    
-}
