+++ Core/Infrastructure/VMD/Interfaces/IBaseVmd.cs
@@ -1,0 +1,5 @@
+using ReactiveUI;
+
+namespace Core.Infrastructure.VMD.Interfaces;
+
+public interface IBaseVmd: IReactiveObject { }
