+++ Core/Infrastructure/VMD/Interfaces/IMenuVmd.cs
@@ -1,0 +1,7 @@
+namespace Core.Infrastructure.VMD.Interfaces;
+
+public interface IMenuVmd<T> : IBaseVmd
+{
+    public IEnumerable<T>? MenuItems { get; init; }
+    
+}
