+++ Core/Stores/SettingsStore.cs
@@ -1,9 +1,9 @@
-using Core.Infrastructure.Models.SettingsModels;
+using Core.Infrastructure.Models.Settings;
 using Core.Infrastructure.Stores;
 
 namespace Core.Stores;
 
-public class SettingsStore : BaseTimerReactiveStore<Settings>
+internal sealed class SettingsStore : BaseTimerReactiveStore<Settings>
 {
     public SettingsStore():base(new Settings()){}
     protected override int InitialTimerSeconds { get; } = 10;
